remove-supabase-client-for-prisma
