+++ src/types/db.d.ts
@@ -1,12 +1,10 @@
 interface ChatMessage {
   id: number; // Primary key
-  chat_id: number;
-  sender_id: number;
-  receiver_id: number;
+  chatId: string;
+  senderId: number;
+  receiverId: number;
   message: string;
-  prev_message_id: number;
-  sent_timestamp: Date;
-  inserted_timestamp: Date;
+  sentTimestamp: Date;
 }
 
 interface Gig {
@@ -21,8 +19,8 @@ interface User {
   address: string;
   username: string | null;
   role: string | null;
-  x_followers: number | null;
-  github_commits: number | null;
-  created_at: string;
+  xFollowers: number | null;
+  githubCommits: number | null;
+  createdAt: string;
   userId: number;
 }
