+++ src/types/pusher.d.ts
@@ -1,4 +1,5 @@
 interface PusherMessage {
   message: string;
   senderId: number;
+  sentiment: string;
 }
